test(popover): migrate dismiss-on-select
